Move TestExecutionStatus from data to api model
