+++ backend/test_observer/controllers/test_executions/test_executions.py
@@ -29,7 +29,6 @@ from test_observer.data_access.models import (
     Stage,
     TestExecution,
 )
-from test_observer.data_access.models_enums import TestExecutionStatus
 from test_observer.data_access.repository import get_or_create
 from test_observer.data_access.setup import get_db
 
@@ -89,9 +88,6 @@ def start_test_execution(
                 "environment_id": environment.id,
                 "artefact_build_id": artefact_build.id,
             },
-            creation_kwargs={
-                "status": TestExecutionStatus.IN_PROGRESS,
-            },
         )
         return {"id": test_execution.id}
     except ValueError as exc:
@@ -107,5 +103,4 @@ def patch_test_execution(
     test_execution = db.query(TestExecution).filter(TestExecution.id == id).one()
     test_execution.c3_link = request.c3_link
     test_execution.jenkins_link = request.jenkins_link
-    test_execution.status = request.status
     db.commit()
